[Garima] rename Icon component to OriginToDestination Component
